feat: 홈 사이드 네비게이션 강의 설정 연동 완료
